requirements_audio.txt
flask==3.0.0
flask-cors==4.0.0
numpy>=1.24.0
pandas>=2.0.0
opensmile-python>=2.5.0
scikit-learn>=1.3.0
joblib>=1.3.0
werkzeug>=3.0.0

